remove user

+++ src/services/chat.service.js
@@ -18,7 +18,7 @@ class ChatService {
 
         await RabbitMQService.sendMessage(room_id, chatMessage);
 
-        return chatMessage 
+        return chatMessage
     }
 
     static createRoom = async (params) => {
@@ -32,9 +32,9 @@ class ChatService {
         }
 
         //Chỉ có trường hợp one-to-one chat mới check exist room
-        if(params.user_ids.length == 2) {
+        if (params.user_ids.length == 2) {
             const checkExistRoom = await RoomRepository.getRoomByUserIDs(params.user_ids)
-            if(checkExistRoom) {
+            if (checkExistRoom) {
                 return RoomRepository.transformForClient(checkExistRoom)
             }
         }
@@ -76,8 +76,8 @@ class ChatService {
     static async getNewMessagesEachRoom(userId) {
         const rooms = await RoomRepository.getRoomsByUserID(userId);
         const roomsTransformed = await RoomRepository.transformForClient(rooms);
-    
-        const messagePromises = rooms.map(room => 
+
+        const messagePromises = rooms.map(room =>
             RedisService.get('newMessage:' + room._id).then(async message => {
                 if (message) {
                     const transformedData = await ChatRepository.transformForClient(JSON.parse(message));
@@ -86,9 +86,9 @@ class ChatService {
                 return null;
             })
         );
-    
+
         const messageResults = await Promise.all(messagePromises);
-    
+
         return roomsTransformed.map((room, index) => ({
             room,
             newMessage: messageResults[index]?.message
@@ -97,18 +97,30 @@ class ChatService {
 
     static async deleteMessage(message_id, userId) {
         const message = await ChatRepository.getMessageById(message_id);
-    
+
         if (!message) {
             throw new NotFoundError("Message not found");
         }
-    
+
         if (message.user_id.toString() !== userId.toString()) {
             throw new BadRequestError("You are not allowed to delete this message");
         }
-    
+
         await ChatRepository.deleteMessage(message_id);
     }
-    
+
+    static async editMessage(message_id, userId, content) {
+        const message = await ChatRepository.getMessageById(message_id);
+        if (!message) {
+            throw new NotFoundError("Message not found");
+        }
+        if (message.user_id.toString() !== userId.toString()) {
+            throw new BadRequestError("You are not allowed to edit this message");
+        }
+
+        await ChatRepository.editMessage(message_id, content, userId);
+    }
+
     static async updateNewMessagesInRoom(roomId, message) {
         const key = 'newMessage:' + roomId;
         await RedisService.set(key, JSON.stringify(message));
@@ -116,26 +128,26 @@ class ChatService {
 
     static async getMessagesInRoom(room_id, page = 1, limit = 10) {
         const skip = (page - 1) * limit;
-    
+
         const [room, messages, totalMessages] = await Promise.all([
             RoomRepository.getRoomByID(room_id),
             ChatRepository.getMessagesByRoomId(room_id, skip, limit),
             ChatRepository.countMessagesByRoomId(room_id)
         ]);
-    
+
         if (!room) {
             throw new NotFoundError("Room not found");
         }
-    
+
         const transformedMessages = await Promise.all(
             messages.map(message => ChatRepository.transformForClient(message))
         );
-    
+
         const totalPages = Math.ceil(totalMessages / limit);
-    
+
         return {
             messages: transformedMessages,
-            currentPage: parseInt(page),    
+            currentPage: parseInt(page),
             totalPages,
             amount: transformedMessages.length,
             totalMessages,
@@ -144,72 +156,123 @@ class ChatService {
         };
     }
 
-    static async addUsersToRoom(room_id, newUserIds, userId) {    
+    static async addUsersToRoom(room_id, newUserIds, userId) {
         let updatedRoom = await RoomRepository.addUsersToRoom(room_id, newUserIds);
-    
+
         if (updatedRoom.user_ids.length > 2) {
             updatedRoom.is_group = true;
         }
-    
+
         const userDetailsPromises = updatedRoom.user_ids.map(findUserById);
         const userDetails = await Promise.all(userDetailsPromises);
-    
+
         if (updatedRoom.is_group && updatedRoom.auto_name) {
             const usersName = userDetails.map(user => user.name);
             updatedRoom.name = usersName.join(', ');
         }
-    
+
         updatedRoom = await RoomRepository.updateRoom(updatedRoom);
-    
+
         await RoomRepository.updateRedisCacheForRoom(updatedRoom);
-    
+
         updatedRoom = await RoomRepository.transformForClient(updatedRoom);
-    
+
         return updatedRoom;
     }
 
     static async removeUsersFromRoom(room_id, user_ids, userId) {
-        let updatedRoom = await RoomRepository.removeUsersFromRoom(room_id, user_ids);
-    
-        if (updatedRoom.user_ids.length > 2) {
-            updatedRoom.is_group = true;
+        const room = await RoomRepository.getRoomByID(room_id);
+        if (!room) {
+            throw new Error('Room not found');
         }
-    
-        const userDetailsPromises = updatedRoom.user_ids.map(findUserById);
-        const userDetails = await Promise.all(userDetailsPromises);
-    
-        if (updatedRoom.is_group && updatedRoom.auto_name) {
-            const usersName = userDetails.map(user => user.name);
-            updatedRoom.name = usersName.join(', ');
+      
+        const isUserInRoom = room.user_ids.includes(userId);
+        if (!isUserInRoom) {
+            throw new Error('User is not in the room');
         }
-    
-        updatedRoom = await RoomRepository.updateRoom(updatedRoom);
-    
-        await RoomRepository.updateRedisCacheForRoom(updatedRoom);
-    
-        updatedRoom = await RoomRepository.transformForClient(updatedRoom);
-    
-        return updatedRoom;
+
+        if (room.is_group && room.created_by.toString() !== userId.toString()) {
+            throw new Error('Only the creator of the group can remove users');
+        }
+        if (user_ids.includes(userId)) {
+            throw new Error('You cannot remove yourself from the room');
+        }
+        if (!room.is_group) {
+            throw new Error('Cannot remove users from a one-to-one chat');
+        }
+        try {
+            let updatedRoom = await RoomRepository.removeUsersFromRoom(room_id, user_ids);
+
+            const userDetailsPromises = updatedRoom.user_ids.map(findUserById);
+            const userDetails = await Promise.all(userDetailsPromises);
+
+            if (updatedRoom.is_group && updatedRoom.auto_name) {
+                const usersName = userDetails.map(user => user.name);
+                updatedRoom.name = usersName.join(', ');
+            }
+            
+            updatedRoom = await RoomRepository.updateRoom(updatedRoom);
+
+            await RoomRepository.updateRedisCacheForRoom(updatedRoom);
+
+            await RoomRepository.deleteListRoomRedisAfterRemoveUser(room_id, user_ids,userId);
+            
+            console.log("update list room user who remove")
+            await RoomRepository.updateRedisCacheForListRoom(room_id, userId);
+
+            updatedRoom = await RoomRepository.transformForClient(updatedRoom);
+            
+            if (updatedRoom.room_user_ids.length === 1) {
+                console.log('delete room')
+                await this.leaveRoom(room_id, userId);
+                return;
+            }
+            return updatedRoom;
+        }
+        catch (error) {
+            throw new BadRequestError(error);
+        }
     }
 
     static async leaveRoom(room_id, userId) {
+        console.log('leave room')
         const room = await RoomRepository.getRoomByID(room_id);
-    
         if (!room) {
-            throw new NotFoundError("Room not found");
+            throw new Error('Room not found');
         }
-    
+      
+        const isUserInRoom = room.user_ids.includes(userId);
+        if (!isUserInRoom) {
+            throw new Error('User is not in the room');
+        }      
+
+        if (!room.is_group) {
+            throw new Error('Cannot leave a one-to-one chat');
+        }
+
+        if (!room.user_ids.includes(userId)) {
+            throw new BadRequestError("User is not in the room");
+        }
+
+        console.log(room.user_ids.length +'delete')
         if (room.user_ids.length === 1) {
-            await RoomRepository.deleteRoom(room_id);
+            await RoomRepository.deleteRoomDb(room_id);
+            console.log('delete room redis')
+            await RoomRepository.deleteListAndRoomRedis(room_id, userId);
             return;
         }
-    
+        
         await RoomRepository.removeUsersFromRoom(room_id, [userId]);
+        
+        await RoomRepository.deleteListRoomRedisAfterRemoveUser(room_id, [userId],userId);
+        
+        await RoomRepository.updateRedisCacheForListRoomLeave(room_id, userId);
         return;
     }
 
-    static async listRooms(userId) {
-        const rooms = await RoomRepository.getRoomsByUserID(userId);
+    static async listRooms(userId, page, limit) {
+        const offset = (page - 1) * limit;
+        const rooms = await RoomRepository.getListRoomsByUserID(userId, offset, limit);
         const roomsTransformed = await RoomRepository.transformForClient(rooms);
         return roomsTransformed;
     }
